requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.13
Flask==2.3.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.9.0
grpc-stubs==1.53.0.6
grpcio==1.74.0
grpcio-status==1.71.2
gTTS==2.4.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
types-protobuf==4.25.0.20240417
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
Werkzeug==3.1.3
